refactor(deriv): hoist 2n+1 second-derivative orchestration to base (Phase 3d)
Move the polarizability, APT (dipole derivatives), and molecular Hessian 2n+1
assemblies into CorrelatedDerivs, closing the property-assembly work of the
refactor. MP2 and CC share one polarizability; the base assemblies are
method-agnostic (MP2 drives all three; CC drives polarizability).

- One perturbed solve yields everything: _perturbed_relaxed_density / _so_
  now return PerturbedResponse(dDrel, dGam, dW). The unrelaxed 2-PDM response
  dGam and the perturbed energy-weighted density dW = d_x I'(Drel, Gam) fall
  out of the same solve, removing MP2's prior 2-3x redundant recomputation
  (separate _perturbed_densities + a _perturbed_lagrangian adapter).
- polarizability / dipole_derivatives / hessian are public on the base (the
  pycc.properties facade reaches them by name via getattr), driven by the
  orbital-response record, the PerturbedResponse hook, wfn.derivatives, and
  perturbed_mo_gauge (threaded into _full_U; non-canonical for MP2, identical
  to the old no-arg call). Route validation lives in the base methods.
- The base is the public correlation-property API; a leaf overrides a property
  only to add method-specific behavior. MP2 adds nothing -> its three wrappers
  are deleted (all inherited). CC keeps a thin polarizability override for its
  model/(T) guards, then defers to super(); _so_polarizability and the inline
  spatial loop are gone (base dispatches SO vs spatial).
- Naming (per PI): dropped the redundant _matrix suffix from the perturbed
  Lagrangian methods and the redundant _2n1 suffix from the assembly names
  (the explicit route was removed in 1c). Public route arguments unchanged.

Zero behavior change. Validated: MP2 test_061/067/068/069 + facade test_074
(52 passed); CC test_076/078/083/084/086 (31 passed) + 2 slow SO HOF
polarizability (AE+FC, 2 passed). SO==spatial holds for MP2 pol/APT/Hessian
and CCSD polarizability.
