+++ pycc/mpderiv.py
@@ -108,38 +108,9 @@ class MPderiv(CorrelatedDerivs):
     # The relaxed-density gradient is already the 2n+1 first derivative; its second derivative
     # (polarizability/APT/Hessian) needs the *response* of the relaxed density, whose new piece
     # is the perturbed Z-vector z^x (same orbital Hessian as the gradient, perturbed RHS).  The
-    # assembly lives in CorrelatedDerivs._{so_}perturbed_relaxed_density, driven by the
-    # _perturbed_unrelaxed_densities hook above.  See mp2_2n1_perturbed.tex / DERIVATIVES_PLAN.
-
-    def _so_perturbed_lagrangian(self, pert, D=None, dD=None) -> np.ndarray:
-        """First-order response ``d_x I'`` of the GSB orbital Lagrangian (``nmo x nmo``),
-        spin-orbital. Differentiates :meth:`_so_lagrangian` term by term with the perturbed
-        integrals (:meth:`CPHF.perturbed_fock`/`perturbed_eri`) and the density response.
-
-        With ``D``/``dD`` omitted this uses the **unrelaxed** 1-PDM and its response
-        (:meth:`_perturbed_densities`) -- the Z-vector RHS derivative. Passing ``D`` = the
-        relaxed 1-PDM and ``dD`` = its response (:meth:`_so_perturbed_relaxed_density`) gives
-        instead ``d_x W`` (the perturbed energy-weighted density; :meth:`_so_gradient` uses
-        ``W = I'(D_rel)`` with the nuclear ``S^X``). The 2-PDM response ``dGam`` is the same
-        either way (:meth:`_perturbed_densities`). The ``termA`` derivative is the full Fock
-        matrix product ``d_x f @ (D + D.T)`` (the diagonal ``d_x eps`` alone suffices for the
-        unrelaxed ``D`` -- no ov block -- but the relaxed ``D`` has ov/core-active blocks that
-        the off-diagonal ``d_x f`` couples).
-
-        MP2 (closed-form) adapter: builds the perturbed integrals + densities and evaluates the
-        shared :meth:`CorrelatedDerivs._so_perturbed_lagrangian_matrix`."""
-        o = self.mp.o
-        ncore = o.stop - self.mp.no
-        cphf = self._full_occ_cphf()
-        Gam = np.asarray(self.mp._so_mp2_tpdm())
-        df = np.asarray(cphf.perturbed_fock(pert, ncore))
-        deri = np.asarray(cphf.perturbed_eri(pert, ncore))
-        dDg, dGam = self._perturbed_densities(pert)
-        dGam = np.asarray(dGam)
-        if D is None:
-            D = self._so_orbital_response().D         # unrelaxed 1-PDM
-            dD = np.asarray(dDg)
-        return self._so_perturbed_lagrangian_matrix(df, deri, D, dD, Gam, dGam)
+    # assembly (and the polarizability/APT/Hessian orchestration) lives in CorrelatedDerivs,
+    # driven by the _perturbed_unrelaxed_densities hook above.  See mp2_2n1_perturbed.tex /
+    # DERIVATIVES_PLAN.
 
     def _unrelaxed_densities(self):
         """MP2 unrelaxed reduced densities as full-MO arrays: the 1-PDM ``D`` (the ``Doo``/``Dvv``
@@ -158,304 +129,9 @@ class MPderiv(CorrelatedDerivs):
         D[v, v] = np.asarray(Dvv)
         return D, np.asarray(Gam)
 
-    def _perturbed_lagrangian(self, pert, D=None, dD=None) -> np.ndarray:
-        """Spatial first-order response ``d_x I'`` of the GSB orbital Lagrangian (``nmo x nmo``),
-        the closed-shell analogue of :meth:`_so_perturbed_lagrangian`: the spin-adapted ``L``
-        (and its derivative ``dL = 2 d_x<pq|rs> - d_x<pq|sr>``) in the two-electron 1-PDM term,
-        ``<pr|st>`` (H.ERI) with ``Gamma`` in the 2-PDM term. ``D``/``dD`` default to the
-        unrelaxed 1-PDM and its response (Z-vector RHS); pass the relaxed 1-PDM and its response
-        for ``d_x W``. See :meth:`_so_perturbed_lagrangian`.
-
-        This is the MP2 (closed-form) adapter: it builds the perturbed integrals ``df``/``deri``/``dL``
-        (CPHF) and the perturbed densities ``dDg``/``dGam`` (:meth:`_perturbed_densities`), then
-        evaluates the shared :meth:`CorrelatedDerivs._perturbed_lagrangian_matrix`."""
-        o = self.mp.o
-        ncore = o.stop - self.mp.no
-        cphf = self._full_occ_cphf()
-        Gam = np.asarray(self.mp._mp2_tpdm())
-        df = np.asarray(cphf.perturbed_fock(pert, ncore))
-        deri = np.asarray(cphf.perturbed_eri(pert, ncore))
-        dL = 2.0 * deri - deri.swapaxes(2, 3)
-        dDg, dGam = self._perturbed_densities(pert)
-        dGam = np.asarray(dGam)
-        if D is None:
-            D = self._orbital_response().D             # unrelaxed 1-PDM
-            dD = np.asarray(dDg)
-        return self._perturbed_lagrangian_matrix(df, deri, dL, D, dD, Gam, dGam)
-
     # ---- second derivatives: polarizability, APT (dipole derivatives), Hessian ----
-
-    def polarizability(self, route: str = '2n+1') -> np.ndarray:
-        """MP2 **correlation** contribution to the static (omega=0) dipole polarizability
-        (a.u.), shape ``(3, 3)``: ``alpha_corr_ab = -d^2 E_corr / dF_a dF_b``, via the 2n+1
-        route (:meth:`_polarizability_2n1`): differentiate the relaxed-density gradient in a
-        field, using only the first-order perturbed *relaxed* density
-        (:meth:`_so_perturbed_relaxed_density` / :meth:`_perturbed_relaxed_density`, which carry the
-        perturbed Z-vector). Frozen-core aware (both spin-orbital and spin-adapted paths).
-
-        ``route`` accepts only ``'2n+1'`` (the sole route; the argument is retained for a uniform
-        property signature). The reference part is kept separate (:meth:`HFwfn.polarizability`);
-        the total is their sum. Electric field only."""
-        if route != '2n+1':
-            raise ValueError(f"unknown polarizability route {route!r} (only '2n+1')")
-        return self._polarizability_2n1()
-
-    def _polarizability_2n1(self) -> np.ndarray:
-        """MP2 correlation polarizability via the 2n+1 route (frozen-core aware; spin-orbital
-        and spin-adapted paths).
-
-        Differentiating the relaxed-density gradient in a field: ``d_b E = -Tr(D_rel mu_b)``
-        (field skeleton ``f^(b) = -mu_b``), so::
-
-            alpha_ab = sum_pq d_a D_rel_pq (mu_b)_pq
-                     + sum_pq D_rel_pq [ (U^a).T mu_b + mu_b U^a ]_pq
-
-        The first term is the perturbed relaxed density (:meth:`_so_perturbed_relaxed_density`,
-        carrying the perturbed Z-vector and, for frozen core, the perturbed core-active divide);
-        the second is the MO dipole rotating under the field (``d_a f^(b) = rotate(U^a, -mu_b)``,
-        with ``U^a`` over the full occupied space -- ``ncore`` canonical core-active block). No
-        second-order CPHF ``U^{ab}`` -- only first-order responses. Reproduces the explicit
-        route to ~machine precision."""
-        from .cphf import Perturbation
-        ncore = self.mp.o.stop - self.mp.no
-        cphf = self._full_occ_cphf()
-        c = self.contract
-        if self.mp.orbital_basis == 'spinorbital':
-            Drel = self._so_orbital_response().Drel
-            perturbed_opdm = self._so_perturbed_relaxed_density
-        else:
-            Drel = self._orbital_response().Drel
-            perturbed_opdm = self._perturbed_relaxed_density
-        mu = [np.asarray(self.mp.H.mu[a]) for a in range(3)]
-        field = [Perturbation('field', b) for b in range(3)]
-        alpha = np.zeros((3, 3))
-        for b in range(3):
-            dDrel = perturbed_opdm(field[b])
-            Ub = np.asarray(cphf._full_U(field[b], ncore))
-            for a in range(3):
-                rot = Ub.T @ mu[a] + mu[a] @ Ub
-                alpha[a, b] = c('pq,pq->', dDrel, mu[a]) + c('pq,pq->', Drel, rot)
-        return alpha
-
-    def dipole_derivatives(self, route: str = '2n+1-field') -> np.ndarray:
-        """MP2 **correlation** contribution to the atomic polar tensors (nuclear dipole
-        derivatives, a.u.), shape ``(natom, 3, 3)`` indexed ``[A, beta, alpha]`` =
-        ``d(mu_alpha)/d(X_{A,beta}) = -d^2 E_corr / dF_alpha dX_{A,beta}`` -- the mixed
-        field/nuclear analog of :meth:`polarizability`.
-
-        ``route='2n+1-field'`` (default) / ``'2n+1-nuclear'`` -- the two 2n+1 routes
-        (:meth:`_dipole_derivatives_2n1`), which use only first-order responses (no ``U^{FX}``).
-        ``'2n+1-nuclear'`` differentiates the relaxed dipole w.r.t. the nuclei (``3N`` nuclear
-        relaxed-density responses); ``'2n+1-field'`` differentiates the relaxed nuclear gradient
-        w.r.t. the field (3 field responses -- ``d_F D_rel``, ``d_F Gamma``, ``d_F W`` -- contracted
-        with the ``3N`` nuclear skeletons). Both give the same tensor; ``'2n+1-field'`` is cheaper
-        and is what the IR/VCD path stores.
-
-        Correlation only; the nuclear ``Z_A`` and SCF reference terms are separate, and the total
-        (nuclear + reference + correlation) is assembled by :func:`pycc.apt`. Basis-aware.
-        Validated against a finite nuclear displacement of the analytic dipole (``test_068``)."""
-        if route not in ('2n+1-nuclear', '2n+1-field'):
-            raise ValueError(f"unknown dipole-derivative route {route!r} "
-                             "(use '2n+1-nuclear' or '2n+1-field')")
-        return self._dipole_derivatives_2n1(route)
-
-    def _dipole_derivatives_2n1(self, route: str) -> np.ndarray:
-        """MP2 correlation atomic polar tensors via the 2n+1 route (both spin paths, frozen-core
-        aware); ``route`` is ``'2n+1-nuclear'`` or ``'2n+1-field'``. See :meth:`dipole_derivatives`.
-
-        Nuclear side -- differentiate the relaxed dipole ``Tr(D_rel mu_a)`` w.r.t. the nucleus
-        (the field gradient has no ``S^X``/2e-skeleton term, so no energy-weighted density
-        appears)::
-
-            P[X,a] = Tr(d_X D_rel mu_a) + Tr(D_rel [mu_a^X + rotate(U^X, mu_a)]).
-
-        Field side -- differentiate the relaxed nuclear gradient
-        ``E^X = sum D_rel f^X + sum Gamma <pq||rs>^X + sum W S^X`` w.r.t. the field::
-
-            P[X,a] = -[ sum d_a D_rel f^X + sum D_rel d_a f^X + sum d_a Gamma <>^X
-                        + sum Gamma d_a <>^X + sum d_a W S^X + sum W d_a S^X ],
-
-        with the 3 field responses ``d_a D_rel`` (:meth:`_so_perturbed_relaxed_density`),
-        ``d_a Gamma`` (:meth:`_perturbed_densities`), and the perturbed energy-weighted density
-        ``d_a W`` (:meth:`_so_perturbed_lagrangian` at the relaxed density). The field-derivatives
-        of the nuclear skeletons carry the orbital rotation ``rotate(U^a, .)`` plus, for
-        ``d_a f^X``, the occupied-sum response and the ``-mu_a^X`` mixed skeleton (the field
-        enters ``h``)."""
-        from .cphf import Perturbation
-        c = self.contract
-        so = self.mp.orbital_basis == 'spinorbital'
-        o = self.mp.o
-        ofull = slice(0, o.stop)
-        ncore = o.stop - self.mp.no
-        cphf = self._full_occ_cphf()
-        d = self.mp.derivatives
-        natom = d.natom
-        Drel = (self._so_orbital_response() if so else self._orbital_response()).Drel
-        popdm = self._so_perturbed_relaxed_density if so else self._perturbed_relaxed_density
-        mu = [np.asarray(self.mp.H.mu[a]) for a in range(3)]
-        P = np.zeros((natom, 3, 3))
-
-        if route == '2n+1-nuclear':
-            for A in range(natom):
-                dip = d.so_dipole(A) if so else d.dipole(A)          # [alpha*3 + beta]
-                for beta in range(3):
-                    pX = Perturbation('nuclear', (A, beta))
-                    dDrel = popdm(pX)
-                    UX = np.asarray(cphf._full_U(pX, ncore))
-                    for alpha in range(3):
-                        dmu = np.asarray(dip[alpha * 3 + beta])       # skeleton d(mu_a)/dX_beta
-                        rot = UX.T @ mu[alpha] + mu[alpha] @ UX
-                        P[A, beta, alpha] = (c('pq,pq->', dDrel, mu[alpha])
-                                             + c('pq,pq->', Drel, dmu + rot))
-            return P
-
-        # route == '2n+1-field'
-        lagr = self._so_perturbed_lagrangian if so else self._perturbed_lagrangian
-        Gam = np.asarray(self.mp._so_mp2_tpdm() if so else self.mp._mp2_tpdm())
-        W = self._lagrangian(Drel, Gam)
-        field = [Perturbation('field', a) for a in range(3)]
-        dDrel = [popdm(field[a]) for a in range(3)]
-        dGamF = [np.asarray(self._perturbed_densities(field[a])[1]) for a in range(3)]
-        dW = [lagr(field[a], Drel, dDrel[a]) for a in range(3)]      # perturbed energy-weighted density
-        U = [np.asarray(cphf._full_U(field[a], ncore)) for a in range(3)]
-
-        def rot1(Um, M):
-            return Um.T @ M + M @ Um
-
-        def rot4(Um, T):
-            return (c('tp,tqrs->pqrs', Um, T) + c('tq,ptrs->pqrs', Um, T)
-                    + c('tr,pqts->pqrs', Um, T) + c('ts,pqrt->pqrs', Um, T))
-
-        for A in range(natom):
-            hx = d.so_core(A) if so else d.core(A)
-            Sx = d.so_overlap(A) if so else d.overlap(A)
-            dip = d.so_dipole(A) if so else d.dipole(A)
-            if so:
-                eriF = [np.asarray(e) for e in d.so_eri(A)]          # <pq||rs>^X (Fock and Gamma)
-                gamX = eriF
-            else:
-                phys = [np.asarray(ch).transpose(0, 2, 1, 3) for ch in d.eri(A)]  # <pq|rs>^X (Gamma)
-                eriF = [2.0 * p - p.transpose(0, 1, 3, 2) for p in phys]          # L^X (Fock)
-                gamX = phys
-            for beta in range(3):
-                fX = np.asarray(hx[beta]) + c('pmqm->pq', eriF[beta][:, ofull, :, ofull])
-                SX = np.asarray(Sx[beta])
-                gX, eX = gamX[beta], eriF[beta]
-                for alpha in range(3):
-                    Um = U[alpha]
-                    muX = np.asarray(dip[alpha * 3 + beta])
-                    occ = (c('rm,prqm->pq', Um[:, ofull], eX[:, :, :, ofull])
-                           + c('rm,pmqr->pq', Um[:, ofull], eX[:, ofull, :, :]))
-                    dfX = rot1(Um, fX) - muX + occ
-                    P[A, beta, alpha] = -(c('pq,pq->', dDrel[alpha], fX) + c('pq,pq->', Drel, dfX)
-                                          + c('pqrs,pqrs->', dGamF[alpha], gX) + c('pqrs,pqrs->', Gam, rot4(Um, gX))
-                                          + c('pq,pq->', dW[alpha], SX) + c('pq,pq->', W, rot1(Um, SX)))
-        return P
-
-    def hessian(self, route: str = '2n+1') -> np.ndarray:
-        """MP2 **correlation** contribution to the molecular (nuclear) Hessian (a.u.), shape
-        ``(3*natom, 3*natom)`` indexed ``(A*3+a, B*3+b)`` = ``d^2 E_corr / dX_{Aa} dX_{Bb}`` --
-        the nuclear-nuclear analog of :meth:`polarizability`/:meth:`dipole_derivatives`, via the
-        2n+1 route (:meth:`_hessian_2n1`): differentiate the relaxed nuclear gradient w.r.t. a
-        second nucleus, using only ``3N`` *first-order* solves (the perturbed relaxed density /
-        energy-weighted density and ``U^Y``). The nuclear-nuclear analog of the ``'2n+1-field'``
-        APT, with the full second skeletons ``f^{XY}``/``<pq||rs>^{XY}``/``S^{XY}``.
-
-        ``route`` accepts only ``'2n+1'`` (retained for a uniform property signature). The
-        reference and nuclear parts are separate; the total (nuclear + reference + correlation) is
-        assembled by :func:`pycc.hessian`. Basis-aware. Validated against a finite difference of the
-        analytic gradient (``test_069``)."""
-        if route != '2n+1':
-            raise ValueError(f"unknown hessian route {route!r} (only '2n+1')")
-        return self._hessian_2n1()
-
-    def _hessian_2n1(self) -> np.ndarray:
-        """MP2 correlation molecular Hessian via the 2n+1 route (both spin paths, frozen-core
-        aware). Differentiate the relaxed nuclear gradient
-        ``E^X = sum D_rel f^X + sum Gamma <pq||rs>^X + sum W S^X`` w.r.t. a second nucleus ``Y``::
-
-            H[X,Y] = sum d_Y D_rel f^X + sum D_rel d_Y f^X + sum d_Y Gamma <>^X
-                     + sum Gamma d_Y <>^X + sum d_Y W S^X + sum W d_Y S^X,
-
-        the nuclear-nuclear analog of the ``'2n+1-field'`` APT (:meth:`_dipole_derivatives_2n1`).
-        Only ``3N`` first-order solves -- the perturbed relaxed density ``d_Y D_rel``
-        (:meth:`_so_perturbed_relaxed_density`), the perturbed energy-weighted density ``d_Y W``
-        (:meth:`_so_perturbed_lagrangian` at ``D_rel``), ``d_Y Gamma``
-        (:meth:`_perturbed_densities`), and ``U^Y`` (:meth:`CPHF._full_U`) -- vs the explicit
-        route's ``O(N^2)`` ``U^{XY}`` solves.
-
-        The field-derivatives of the nuclear skeletons carry (i) the full second integral
-        skeletons ``f^{XY}``/``<>^{XY}``/``S^{XY}`` (:meth:`CPHF._d2int_blocks`, cached per atom
-        pair -- all nonzero here, unlike the field case where only ``-mu^X`` survived), and (ii)
-        the ``U^Y`` orbital rotation of the ``X`` skeletons. The rotations are hoisted off the
-        ``O(N^2)`` pair loop onto the (per-``Y``) densities via ``sum A rot(U,B) = sum rot(U^T,A) B``:
-        ``Dtil = U D + D U^T``, ``Wtil`` likewise, ``Gtil = rotate4(U^T, Gamma)``, and the
-        Fock skeleton's occupied-sum response as the per-``X`` intermediate ``J^X`` contracted
-        with ``U^Y`` (so no ``O(N^2)`` four-index rotation)."""
-        from .cphf import Perturbation
-        c = self.contract
-        so = self.mp.orbital_basis == 'spinorbital'
-        ofull = slice(0, self.mp.o.stop)
-        ncore = self.mp.o.stop - self.mp.no
-        cphf = self._full_occ_cphf()
-        d = self.mp.derivatives
-        natom = d.natom
-        nc = 3 * natom
-        Drel = (self._so_orbital_response() if so else self._orbital_response()).Drel
-        Gam = np.asarray(self.mp._so_mp2_tpdm() if so else self.mp._mp2_tpdm())
-        W = self._lagrangian(Drel, Gam)
-        popdm = self._so_perturbed_relaxed_density if so else self._perturbed_relaxed_density
-        lagr = self._so_perturbed_lagrangian if so else self._perturbed_lagrangian
-        pert = [Perturbation('nuclear', (A, ct)) for A in range(natom) for ct in range(3)]
-
-        def rot4(Um, T):
-            return (c('tp,tqrs->pqrs', Um, T) + c('tq,ptrs->pqrs', Um, T)
-                    + c('tr,pqts->pqrs', Um, T) + c('ts,pqrt->pqrs', Um, T))
-
-        # first-order responses + hoisted per-Y rotated densities (sum A rot(U,B) = sum rot(U^T,A) B)
-        dDrel = [popdm(p) for p in pert]
-        dGamN = [np.asarray(self._perturbed_densities(p)[1]) for p in pert]
-        dW = [lagr(pert[i], Drel, dDrel[i]) for i in range(nc)]
-        U = [np.asarray(cphf._full_U(p, ncore)) for p in pert]
-        Dtil = [U[i] @ Drel + Drel @ U[i].T for i in range(nc)]
-        Wtil = [U[i] @ W + W @ U[i].T for i in range(nc)]
-        Gtil = [rot4(U[i].T, Gam) for i in range(nc)]
-
-        # per-X first skeletons; J^X carries the Fock skeleton's occupied-sum rotation response
-        fX, gamX, SX, JX = [], [], [], []
-        for p in pert:
-            A, ct = p.comp
-            hx = np.asarray((d.so_core(A) if so else d.core(A))[ct])
-            Sx = np.asarray((d.so_overlap(A) if so else d.overlap(A))[ct])
-            if so:
-                eF = np.asarray(d.so_eri(A)[ct])
-                gm = eF
-            else:
-                ph = np.asarray(d.eri(A)[ct]).transpose(0, 2, 1, 3)     # <pq|rs>^X (Gamma)
-                eF = 2.0 * ph - ph.transpose(0, 1, 3, 2)                # L^X (Fock)
-                gm = ph
-            fX.append(hx + c('pmqm->pq', eF[:, ofull, :, ofull]))
-            gamX.append(gm)
-            SX.append(Sx)
-            JX.append(c('pq,prqm->rm', Drel, eF[:, :, :, ofull])
-                      + c('pq,pmqr->rm', Drel, eF[:, ofull, :, :]))
-
-        H = np.zeros((nc, nc))
-        for iy, py in enumerate(pert):
-            Ay, cy = py.comp
-            for ix, px in enumerate(pert):
-                Ax, cx = px.comp
-                blk = cphf._d2int_blocks(Ax, Ay)             # raw second skeletons (no U^{XY})
-                core2 = blk['core'][cx * 3 + cy]
-                ov2 = blk['overlap'][cx * 3 + cy]
-                e2 = blk['eri'][cx * 3 + cy]
-                L2 = e2 if so else 2.0 * e2 - e2.swapaxes(2, 3)
-                f2 = core2 + c('pmqm->pq', L2[:, ofull, :, ofull])       # f^{XY}
-                H[ix, iy] = (c('pq,pq->', dDrel[iy] + Dtil[iy], fX[ix]) + c('pq,pq->', Drel, f2)
-                             + c('pqrs,pqrs->', dGamN[iy] + Gtil[iy], gamX[ix]) + c('pqrs,pqrs->', Gam, e2)
-                             + c('pq,pq->', dW[iy] + Wtil[iy], SX[ix]) + c('pq,pq->', W, ov2)
-                             + float(np.sum(U[iy][:, ofull] * JX[ix])))
-        return H
+    # Inherited unchanged from CorrelatedDerivs (polarizability / dipole_derivatives / hessian);
+    # MP2 adds no method-specific handling, so there is no override here.
 
     # ---- atomic axial tensors (VCD, magnetic/nuclear mixed derivative) ----
